feat(backend): Create agents controller.

+++ src/main/java/com/helper/vavahelper/Repositories/AgentsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.helper.vavahelper.models.Agents.Agents;
 
-public interface AgentsRepository extends JpaRepository<Agents, Integer>{
+public interface AgentsRepository extends JpaRepository<Agents, Long>{
     
 }
